requirements.txt
# Core dependencies
requests>=2.31.0
psycopg2-binary>=2.9.9
sshtunnel>=0.4.0
paramiko>=2.12.0,<3.0.0  # sshtunnel 0.4.0 requires paramiko 2.x
python-dotenv>=1.0.0

# Outbound lead generation (agent-driven)
anthropic>=0.40.0

# Testing (optional, for development)
# pytest>=7.4.0
# pytest-cov>=4.1.0

# Type hints (optional, for development)
# mypy>=1.5.0

# Linting (optional, for development)
# ruff>=0.1.0
